+++ src/models/mod.rs
@@ -1,5 +1,6 @@
 pub mod database;
 pub mod operator;
 pub mod operator_repository;
+pub mod session_store;
 pub mod vulnerability;
 pub mod vulnerability_repository;